added new option: -q
Used to load bitstreams other than the MEGA65 bitstream. With "-q" the M65 tool will quit after having loaded the bitstream. This makes sense for non-MEGA65 bitstreams because the M65 tools expects a certain serial protocol to be able to communicate with the MEGA65 which is something that other cores such as the zxuno4mega65 core or the gbc4mega65 core are not offering.