#60: Various efforts to get xemu and mega65_ftp talking nicely with one another.
The main victory this iteration was to add a pause after the remotesd-side sees the $C000 value change due to mega65_ftp's "sc000 1" command.

    if (PEEK(0xC000)) {
      // Perform one or more jobs
      job_count = PEEK(0xC000);

      // pause a little here, to permit xemu to echo the last command back to mega65_ftp
      for (aa = 0; aa < 30000; aa++)
        continue;

The extra pause here prevents the possibility of remotesd-side delivering the sector contents over serial at the same time xemu tries to echo back the "sc000 1" command.

Future efforts could try fix this on the xemu-side, by making it behave the same way as the hardware's $D680 flag for assessing if the sd-card is busy.

If a small delay is created by bringing such a flag high-then-low, then I won't need to add my dummy delay into remotesd, just for xemu's sake.

I'll aim to remove a lot of un-needed debug stuff later on before the pull-request.