I added an 'out.txt' file containing a disassembly of the last known compiled output of 'remotesd.c' that I had used for m65dbg's copy of mega65_ftp.
Sadly, I was unable to track the original 'remotesd.c' source it was generated from, as I had made tweaks to it in order to get it talking to xemu nicely via a 'xemu-flag' check.

Thankfully, I've added enough annotations into the 'out.txt' file to reverse-engineer the changes I had made.

I'm preserving that reverse-engineering effort here, in-case I need to revisit it, or improve on it.