feat: add basic regression test script #124
adds a basic regression test script that uses the two conforming tests.
Also changes the format m65 outputs the log files to a better parseable
format and allows for running -b and -u in tandem by adding a 4 second
delay after bitstream delivery so the system can start.

It also adds the extraction of the running bitstream version for
logging with the unit test.