fix(mega65_ftp): added detection of program in memory #144
This is to prevent starting mega65_ftp helper while in flasher,
as this may result in flashing slot 0!

will check for program in memory after helper was not detected.
if program is found, mega65_ftp will warn and exit.
Option -F was added to force pushing, regardless of program in memory.